+++ DTCD-Visualization-List/src/Plugin.js
@@ -221,4 +221,13 @@ export class VisualizationList extends PanelPlugin {
     };
   }
 
+  getState() {
+    return this.getPluginConfig();
+  }
+
+  setState(newState) {
+    if (typeof newState !== 'object' ) return;
+
+    this.setPluginConfig(newState);
+  }
 }